requirements.txt
openai
fastapi
pandas
ipykernel
python-dotenv
semantic-kernel
graphrag
pdfplumber
uvicorn
python-multipart
quart
crewai